Move _get_neigbor_pairs to utility module.
