+++ emle/models/_mace.py
@@ -30,10 +30,11 @@ __all__ = ["MACEEMLE"]
 import os as _os
 import torch as _torch
 
-from torch import Tensor
-from typing import Optional, Tuple, List
+from typing import List
 
 from ._emle import EMLE as _EMLE
+from ._emle import _has_nnpops
+from ._utils import _get_neighbor_pairs
 
 try:
     import mace.tools as _mace_tools
@@ -49,13 +50,6 @@ try:
     _has_e3nn = True
 except:
     _has_e3nn = False
-
-try:
-    import NNPOps.neighbors.getNeighborPairs as _getNeighborPairs
-
-    _has_nnpops = True
-except:
-    _has_nnpops = False
 
 
 class MACEEMLE(_torch.nn.Module):
@@ -224,6 +218,9 @@ class MACEEMLE(_torch.nn.Module):
         self.register_buffer(
             "_cell", _torch.zeros((3, 3), dtype=self._dtype, requires_grad=False)
         )
+
+        # Set the _get_neighbor_pairs method on the instance.
+        self._get_neighbor_pairs = _get_neighbor_pairs
 
     @staticmethod
     def _to_one_hot(indices: _torch.Tensor, num_classes: int) -> _torch.Tensor:
@@ -417,58 +414,3 @@ class MACEEMLE(_torch.nn.Module):
 
         # Return the MACE and EMLE energy components.
         return _torch.stack([E_vac, E_emle[0], E_emle[1]])
-
-    @staticmethod
-    def _get_neighbor_pairs(
-        positions: _torch.Tensor,
-        cell: Optional[_torch.Tensor],
-        cutoff: float,
-        dtype: _torch.dtype,
-        device: _torch.device,
-    ) -> Tuple[_torch.Tensor, _torch.Tensor]:
-        """
-        Get the shifts and edge indices.
-
-        Notes
-        -----
-        This method calculates the shifts and edge indices by determining neighbor pairs (``neighbors``)
-        and respective wrapped distances (``wrappedDeltas``) using ``NNPOps.neighbors.getNeighborPairs``.
-        After obtaining the ``neighbors`` and ``wrappedDeltas``, the pairs with negative indices (r>cutoff)
-        are filtered out, and the edge indices and shifts are finally calculated.
-
-        Parameters
-        ----------
-        positions : _torch.Tensor
-            The positions of the atoms.
-        cell : _torch.Tensor
-            The cell vectors.
-        cutoff : float
-            The cutoff distance.
-        dtype : _torch.dtype
-            The data type.
-        device : _torch.device
-            The device.
-
-        Returns
-        -------
-        edgeIndex : _torch.Tensor
-            The edge indices.
-        shifts : _torch.Tensor
-            The shifts.
-        """
-        # Get the neighbor pairs, shifts and edge indices.
-        neighbors, wrapped_deltas, _, _ = _getNeighborPairs(positions, cutoff, -1, cell)
-        mask = neighbors >= 0
-        neighbors = neighbors[mask].view(2, -1)
-        wrapped_deltas = wrapped_deltas[mask[0], :]
-
-        edge_index = _torch.hstack((neighbors, neighbors.flip(0))).to(_torch.int64)
-        if cell is not None:
-            deltas = positions[edge_index[0]] - positions[edge_index[1]]
-            wrapped_deltas = _torch.vstack((wrapped_deltas, -wrapped_deltas))
-            shifts_idx = _torch.mm(deltas - wrapped_deltas, _torch.linalg.inv(cell))
-            shifts = _torch.mm(shifts_idx, cell)
-        else:
-            shifts = _torch.zeros((edge_index.shape[1], 3), dtype=dtype, device=device)
-
-        return edge_index, shifts
